refactor: split EndpointHandler into CRUD, query, and publish handlers
Follows Clean Architecture SRP — each handler serves a single actor:
- EndpointHandler: CRUD + archive (admin)
- QueryEndpointHandler: RAG pipeline + policy enforcement (end user)
- PublishEndpointHandler: marketplace publish/unpublish/sync/health

Dependency inversion via callbacks (DeletionCheck, MetadataEnricher)
keeps payment imports out of the endpoints component.
