+++ backend/syft_space/components/endpoints/routes.py
@@ -6,6 +6,8 @@ from fastapi.responses import JSONResponse
 from syft_space.components.auth.dependencies import get_verified_user_email
 from syft_space.components.auth.public import public_route
 from syft_space.components.endpoints.handlers import EndpointHandler
+from syft_space.components.endpoints.publish_handler import PublishEndpointHandler
+from syft_space.components.endpoints.query_handler import QueryEndpointHandler
 from syft_space.components.endpoints.schemas import (
     AuthenticatedQueryRequest,
     CreateEndpointRequest,
@@ -26,44 +28,35 @@ from syft_space.components.tenants.dependency import get_tenant_dependency
 from syft_space.components.tenants.entities import Tenant
 
 
-def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
-    """Build the endpoint routes.
-
-    Args:
-        handler: Endpoint handler instance
-
-    Returns:
-        Configured API router
-    """
+def build_endpoint_routes(
+    handler: EndpointHandler,
+    query_handler: QueryEndpointHandler,
+    publish_handler: PublishEndpointHandler,
+) -> APIRouter:
+    """Build the endpoint routes."""
     router = APIRouter(prefix="/endpoints", tags=["endpoints"])
 
     def get_handler() -> EndpointHandler:
-        """Dependency to get the endpoint handler."""
         return handler
+
+    def get_query_handler() -> QueryEndpointHandler:
+        return query_handler
+
+    def get_publish_handler() -> PublishEndpointHandler:
+        return publish_handler
 
     async def get_verified_sender_email(
         request: Request,
         tenant: Tenant = Depends(get_tenant_dependency),
-        handler: EndpointHandler = Depends(get_handler),
+        handler: PublishEndpointHandler = Depends(get_publish_handler),
     ) -> str:
-        """Get verified sender email from SyftHub token.
-
-        Args:
-            request: FastAPI request object
-            tenant: Current tenant (injected)
-            handler: Endpoint handler (injected)
-
-        Returns:
-            Verified sender email from SyftHub token
-
-        Raises:
-            HTTPException: If no marketplace configured or token invalid
-        """
+        """Get verified sender email from SyftHub token."""
         marketplace = await handler.marketplace_repository.get_default(tenant.id)
         if not marketplace:
             raise HTTPException(status_code=400, detail="No marketplace configured")
-
         return await get_verified_user_email(request, marketplace)
+
+    # ── CRUD routes (EndpointHandler) ────────────────────────────
 
     @router.post("/", response_model=EndpointCreateResponse, status_code=201)
     async def create_endpoint(
@@ -71,15 +64,6 @@ def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
         tenant: Tenant = Depends(get_tenant_dependency),
         handler: EndpointHandler = Depends(get_handler),
     ) -> EndpointCreateResponse:
-        """Create a new endpoint.
-
-        Args:
-            request: Endpoint creation request
-            tenant: Current tenant (injected)
-
-        Returns:
-            Created endpoint details
-        """
         return await handler.create_endpoint(request, tenant)
 
     @router.get("/", response_model=list[EndpointListItem])
@@ -87,42 +71,7 @@ def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
         tenant: Tenant = Depends(get_tenant_dependency),
         handler: EndpointHandler = Depends(get_handler),
     ) -> list[EndpointListItem]:
-        """List all endpoints.
-
-        Args:
-            tenant: Current tenant (injected)
-
-        Returns:
-            List of endpoints with summary information
-        """
         return await handler.list_endpoints(tenant)
-
-    @router.post("/validate-slug", response_model=SlugAvailabilityResponse)
-    async def validate_slug(
-        request: SlugAvailabilityRequest,
-        tenant: Tenant = Depends(get_tenant_dependency),
-        handler: EndpointHandler = Depends(get_handler),
-    ) -> SlugAvailabilityResponse:
-        """Validate if a slug is available locally and optionally on marketplaces.
-
-        This endpoint allows checking slug uniqueness before creating an endpoint.
-        - If neither marketplace_ids nor check_all_marketplaces is provided, only checks locally (fast)
-        - If check_all_marketplaces is True, checks all active marketplaces
-        - If marketplace_ids is provided, checks only those specific marketplaces
-
-        Args:
-            request: Slug and optional marketplace options
-            tenant: Current tenant (injected)
-
-        Returns:
-            Availability status for local and each requested marketplace
-        """
-        return await handler.check_slug_availability(
-            request.slug,
-            request.marketplace_ids,
-            request.check_all_marketplaces,
-            tenant,
-        )
 
     @router.get("/{slug}", response_model=EndpointDetailResponse)
     async def get_endpoint(
@@ -130,15 +79,6 @@ def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
         tenant: Tenant = Depends(get_tenant_dependency),
         handler: EndpointHandler = Depends(get_handler),
     ) -> EndpointDetailResponse:
-        """Get details of a specific endpoint.
-
-        Args:
-            slug: Endpoint slug
-            tenant: Current tenant (injected)
-
-        Returns:
-            Endpoint details
-        """
         return await handler.get_endpoint(slug, tenant)
 
     @router.patch("/{slug}", response_model=EndpointDetailResponse)
@@ -148,19 +88,33 @@ def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
         tenant: Tenant = Depends(get_tenant_dependency),
         handler: EndpointHandler = Depends(get_handler),
     ) -> EndpointDetailResponse:
-        """Update an endpoint's metadata.
-
-        Allows partial updates to name, summary, and description.
-
-        Args:
-            slug: Endpoint slug
-            request: Update request with fields to update
-            tenant: Current tenant (injected)
-
-        Returns:
-            Updated endpoint details
-        """
         return await handler.update_endpoint(slug, request, tenant)
+
+    @router.post("/{slug}/archive", response_model=EndpointDetailResponse)
+    async def archive_endpoint(
+        slug: str,
+        tenant: Tenant = Depends(get_tenant_dependency),
+        handler: EndpointHandler = Depends(get_handler),
+    ) -> EndpointDetailResponse:
+        return await handler.archive_endpoint(slug, tenant)
+
+    @router.post("/{slug}/unarchive", response_model=EndpointDetailResponse)
+    async def unarchive_endpoint(
+        slug: str,
+        tenant: Tenant = Depends(get_tenant_dependency),
+        handler: EndpointHandler = Depends(get_handler),
+    ) -> EndpointDetailResponse:
+        return await handler.unarchive_endpoint(slug, tenant)
+
+    @router.delete("/{slug}", response_model=dict[str, str])
+    async def delete_endpoint(
+        slug: str,
+        tenant: Tenant = Depends(get_tenant_dependency),
+        handler: EndpointHandler = Depends(get_handler),
+    ) -> dict[str, str]:
+        return await handler.delete_endpoint(slug, tenant)
+
+    # ── Query route (QueryEndpointHandler) ───────────────────────
 
     @public_route
     @router.post("/{slug}/query", response_model=QueryEndpointResponse)
@@ -169,29 +123,10 @@ def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
         request: QueryEndpointRequest,
         tenant: Tenant = Depends(get_tenant_dependency),
         sender_email: str = Depends(get_verified_sender_email),
-        handler: EndpointHandler = Depends(get_handler),
+        handler: QueryEndpointHandler = Depends(get_query_handler),
         x_payment: str | None = Header(None, alias="X-Payment", max_length=10000),
     ) -> QueryEndpointResponse | JSONResponse:
-        """Query an endpoint - main RAG flow (PUBLIC, requires SyftHub token).
-
-        This is the core endpoint that orchestrates:
-        - Dataset search (if configured)
-        - Model chat (if configured)
-        - Policy enforcement (pre/post hooks)
-        - MPP payment flow (X-Payment header / HTTP 402)
-
-        Args:
-            slug: Endpoint slug
-            request: Query request with messages and parameters
-            tenant: Current tenant (injected)
-            sender_email: Verified sender email from SyftHub token (injected)
-            handler: Endpoint handler (injected)
-            x_payment: MPP payment credential from X-Payment header (optional)
-
-        Returns:
-            Query response with summary and/or references, or 402 if payment required
-        """
-        # Create authenticated request with verified sender email
+        """Query an endpoint - main RAG flow (PUBLIC, requires SyftHub token)."""
         auth_request = AuthenticatedQueryRequest.from_request(request, sender_email)
 
         try:
@@ -205,7 +140,6 @@ def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
                 headers={"WWW-Authenticate": e.www_authenticate},
             )
 
-        # If there's a payment receipt, include it in the response header
         if payment_receipt:
             return JSONResponse(
                 status_code=200,
@@ -215,64 +149,38 @@ def build_endpoint_routes(handler: EndpointHandler) -> APIRouter:
 
         return response
 
+    # ── Publish routes (PublishEndpointHandler) ──────────────────
+
+    @router.post("/validate-slug", response_model=SlugAvailabilityResponse)
+    async def validate_slug(
+        request: SlugAvailabilityRequest,
+        tenant: Tenant = Depends(get_tenant_dependency),
+        handler: PublishEndpointHandler = Depends(get_publish_handler),
+    ) -> SlugAvailabilityResponse:
+        return await handler.check_slug_availability(
+            request.slug,
+            request.marketplace_ids,
+            request.check_all_marketplaces,
+            tenant,
+        )
+
     @router.post("/{slug}/publish", response_model=PublishEndpointResponse)
     async def publish_endpoint(
         slug: str,
         request: PublishEndpointRequest,
         tenant: Tenant = Depends(get_tenant_dependency),
-        handler: EndpointHandler = Depends(get_handler),
+        handler: PublishEndpointHandler = Depends(get_publish_handler),
     ) -> PublishEndpointResponse:
-        """Publish an endpoint to one or more marketplaces.
-
-        Args:
-            slug: Endpoint slug
-            request: Publish request with marketplace IDs or publish_to_all_marketplaces flag
-            tenant: Current tenant (injected)
-
-        Returns:
-            Publish results for each marketplace
-        """
         return await handler.publish_endpoint(
-            slug,
-            request.marketplace_ids,
-            request.publish_to_all_marketplaces,
-            tenant,
+            slug, request.marketplace_ids, request.publish_to_all_marketplaces, tenant
         )
 
     @router.delete("/{slug}/unpublish", response_model=list[UnpublishResult])
     async def unpublish_endpoint(
         slug: str,
         tenant: Tenant = Depends(get_tenant_dependency),
-        handler: EndpointHandler = Depends(get_handler),
+        handler: PublishEndpointHandler = Depends(get_publish_handler),
     ) -> list[UnpublishResult]:
-        """Unpublish an endpoint.
-        Args:
-            slug: Endpoint slug
-            tenant: Current tenant (injected)
-            handler: Endpoint handler (injected)
-        Returns:
-            Unpublish results for each marketplace
-        Raises:
-            HTTPException: If endpoint not found or marketplace not found
-            HTTPException: If endpoint is not published
-        """
         return await handler.unpublish_endpoint(slug, tenant)
-
-    @router.delete("/{slug}", response_model=dict[str, str])
-    async def delete_endpoint(
-        slug: str,
-        tenant: Tenant = Depends(get_tenant_dependency),
-        handler: EndpointHandler = Depends(get_handler),
-    ) -> dict[str, str]:
-        """Delete an endpoint.
-
-        Args:
-            slug: Endpoint slug
-            tenant: Current tenant (injected)
-
-        Returns:
-            Success message
-        """
-        return await handler.delete_endpoint(slug, tenant)
 
     return router
